Basic support for driving a display with rpi_ws281x
NOTE: weatherscene.py is known to NOT work under Python 3.x due to
an incompatibility with MicroPython's handling of strings and bytes.

This can be easily resolved although I opted to not do this, to preserve
compabilitity with the original code.

To use this on a Raspberry Pi, try:

    sudo apt install -y python-pip python-requests
    sudo pip install rpi_ws281x

Then connect the display's data line to the Raspberry Pi's GPIO 18 (PCM CLK)
(see https://pinout.xyz/)

References:

- https://github.com/noahwilliamsson/lamatrix/issues/1
- https://github.com/rpi-ws281x/rpi-ws281x-python (userspace WS281x driver)
- https://github.com/jgarff/rpi_ws281x (wiring docs)